@@ -1,7 +1,9 @@
 source 'https://rubygems.org'
 
-gem 'rails', '4.2.6'
-gem 'sqlite3'
+ruby '2.3.1'
+
+gem 'rails', '~> 5.0.0'
+gem 'dotenv-rails', require: 'dotenv/rails-now'
 gem 'uglifier', '>= 1.3.0'
 gem 'jquery-rails'
 gem 'paypal-recurring',
@@ -10,4 +12,10 @@ gem 'paypal-recurring',
 
 group :development, :test do
   gem 'rspec-rails', '3.1.0'
+  gem 'sqlite3'
+end
+
+group :production do
+  gem 'pg'
+  gem 'rails_12factor'
 end
