@@ -1,4 +1,4 @@
-class Subscription < ActiveRecord::Base
+class Subscription < ApplicationRecord
   attr_accessor :paypal_payment_token
 
   belongs_to :plan
@@ -6,11 +6,16 @@ class Subscription < ActiveRecord::Base
   delegate :price, :paypal_description, to: :plan
 
   def cancel!
-    update(cancel: true)
+    update(canceled: true)
   end
 
   def paid?
     return false if paid_until.blank?
     paid_until >= Time.current
+  end
+
+  def trial?
+    return false if paid_until.blank?
+    paid_until < created_at
   end
 end
