@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,15 +10,16 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140928002705) do
+ActiveRecord::Schema.define(version: 20161105170303) do
 
   create_table "plans", force: :cascade do |t|
-    t.string   "name",               null: false
-    t.float    "price",              null: false
+    t.string   "name",                           null: false
+    t.float    "price",                          null: false
     t.string   "paypal_description"
     t.text     "description"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.integer  "trial_length",       default: 0, null: false
   end
 
   create_table "subscriptions", force: :cascade do |t|
@@ -29,8 +29,8 @@ ActiveRecord::Schema.define(version: 20140928002705) do
     t.integer  "plan_id"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.boolean  "canceled",          default: false, null: false
+    t.index ["plan_id"], name: "index_subscriptions_on_plan_id"
   end
-
-  add_index "subscriptions", ["plan_id"], name: "index_subscriptions_on_plan_id"
 
 end
