@@ -7,17 +7,23 @@
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 ActiveRecord::Base.transaction do
   Plan.create([
-    {
-      name: 'Mini plan',
-      description: 'Mini plan allows you the mini plan things.',
-      paypal_description: 'Mini plan allows you the mini plan things.',
-      price: 10.50
-    },
-    {
-      name: 'Super plan',
-      description: 'Super plan allows you the mini plan things.',
-      paypal_description: 'Super plan allows you the mini plan things.',
-      price: 30.50
-    }
-  ])
+                {
+                  name: 'Stupid plan',
+                  description: 'Stupid plan is just for stupids.',
+                  paypal_description: 'Stupid plan is just for stupids.',
+                  price: 0.50
+                },
+                {
+                  name: 'Mini plan',
+                  description: 'Mini plan allows you some things.',
+                  paypal_description: 'Mini plan allows you some things.',
+                  price: 10.50
+                },
+                {
+                  name: 'Super plan',
+                  description: 'Super plan is powerful.',
+                  paypal_description: 'Super plan is powerful.',
+                  price: 30.50
+                }
+              ])
 end
